Moves logic from ChatManager to DiscordHook Updates javadocs at DiscordHook
