+++ src/main/java/net/sacredlabyrinth/phaed/simpleclans/managers/ChatManager.java
@@ -3,13 +3,6 @@ package net.sacredlabyrinth.phaed.simpleclans.managers;
 import github.scarsz.discordsrv.DiscordSRV;
 import github.scarsz.discordsrv.api.Subscribe;
 import github.scarsz.discordsrv.api.events.DiscordReadyEvent;
-import github.scarsz.discordsrv.dependencies.jda.api.Permission;
-import github.scarsz.discordsrv.dependencies.jda.api.entities.Category;
-import github.scarsz.discordsrv.dependencies.jda.api.entities.Guild;
-import github.scarsz.discordsrv.dependencies.jda.api.entities.GuildChannel;
-import github.scarsz.discordsrv.dependencies.jda.api.entities.TextChannel;
-import github.scarsz.discordsrv.dependencies.jda.api.requests.restaction.ChannelAction;
-import github.scarsz.discordsrv.objects.managers.AccountLinkManager;
 import me.clip.placeholderapi.PlaceholderAPI;
 import net.sacredlabyrinth.phaed.simpleclans.Clan;
 import net.sacredlabyrinth.phaed.simpleclans.ClanPlayer;
@@ -17,14 +10,13 @@ import net.sacredlabyrinth.phaed.simpleclans.Helper;
 import net.sacredlabyrinth.phaed.simpleclans.SimpleClans;
 import net.sacredlabyrinth.phaed.simpleclans.chat.ChatHandler;
 import net.sacredlabyrinth.phaed.simpleclans.chat.SCMessage;
+import net.sacredlabyrinth.phaed.simpleclans.hooks.discord.DiscordHook;
 import net.sacredlabyrinth.phaed.simpleclans.utils.ChatUtils;
 import org.bukkit.Bukkit;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.*;
-import java.util.concurrent.ExecutionException;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
 
@@ -32,125 +24,27 @@ import static net.sacredlabyrinth.phaed.simpleclans.ClanPlayer.Channel;
 import static net.sacredlabyrinth.phaed.simpleclans.chat.SCMessage.Source;
 import static net.sacredlabyrinth.phaed.simpleclans.managers.SettingsManager.ConfigField;
 import static net.sacredlabyrinth.phaed.simpleclans.managers.SettingsManager.ConfigField.*;
+import static org.bukkit.Bukkit.getServer;
 
-@SuppressWarnings({"ResultOfMethodCallIgnored", "UnusedReturnValue"})
 public final class ChatManager {
 
     private final SimpleClans plugin;
-    private final SettingsManager settingsManager;
+
+    private DiscordHook discordHook;
     private final Set<ChatHandler> handlers = new HashSet<>();
-
-    private Guild guild;
-    private List<String> textCategories;
-
-    private static final int MAX_CHANNELS_PER_CATEGORY = 50;
 
     public ChatManager(SimpleClans plugin) {
         this.plugin = plugin;
-        this.settingsManager = plugin.getSettingsManager();
         registerHandlers();
     }
 
     @Subscribe
-    public void setupDiscord(DiscordReadyEvent event) {
-        guild = DiscordSRV.getPlugin().getMainGuild();
-        textCategories = settingsManager.getStringList(DISCORDCHAT_TEXT_CATEGORY_IDS).stream().
-                filter(this::categoryExists).collect(Collectors.toList());
-
-        Iterator<String> tagIter = getClanTags().iterator();
-        while (tagIter.hasNext()) {
-            if (createChannel(tagIter.next())) {
-                tagIter.remove();
-            } else {
-                break;
-            }
-        }
+    public void registerDiscord(DiscordReadyEvent event) {
+        discordHook = new DiscordHook(plugin);
+        DiscordSRV.api.subscribe(discordHook);
+        getServer().getPluginManager().registerEvents(discordHook, plugin);
     }
 
-    public List<Category> getCategories() {
-        return textCategories.stream().
-                filter(this::categoryExists).
-                map(category -> guild.getCategoryById(category)).
-                collect(Collectors.toList());
-    }
-
-    /**
-     * Creates a new {@link Channel} in available categories,
-     * otherwise create one
-     * <p>
-     * Note: Makes a new category ({@link ChatManager#createCategory()}) if no one available.
-     * </p>
-     *
-     * @param clanTag The clan tag
-     */
-    public boolean createChannel(@NotNull String clanTag) {
-        Category availableCategory = getCategories().stream().
-                filter(category -> category.getTextChannels().size() < MAX_CHANNELS_PER_CATEGORY).
-                findAny().orElseGet(this::createCategory);
-
-        return availableCategory != null && createChannel(availableCategory, clanTag);
-    }
-
-    /**
-     * Creates a new {@link Channel} in defined {@link Category}
-     *
-     * @param category The category, where channel will be created
-     * @param clanTag  The clan tag
-     * @return true, if channel was created. False, if category reached the limit.
-     */
-    public boolean createChannel(@NotNull Category category, @NotNull String clanTag) {
-        ChannelAction<TextChannel> action = category.createTextChannel(clanTag);
-
-        if (category.getTextChannels().size() >= MAX_CHANNELS_PER_CATEGORY) {
-            return false;
-        }
-
-        for (Long discordId : getDiscordPlayersId(clanTag)) {
-            action.addMemberPermissionOverride(discordId,
-                    Collections.singletonList(Permission.VIEW_CHANNEL),
-                    Collections.emptyList());
-        }
-        action.complete();
-        return true;
-    }
-
-    /**
-     * Checks if category can be obtained from id.
-     */
-    public boolean categoryExists(String categoryId) {
-        return guild.getCategoryById(categoryId) != null;
-    }
-
-    /**
-     * Creates a new {@link Category} with numbering
-     *
-     * @return Category
-     */
-    @Nullable
-    public Category createCategory() {
-        String categoryNumeric = String.valueOf(textCategories.size() == 0 ? "" : textCategories.size());
-        String categoryName = settingsManager.getString(DISCORDCHAT_TEXT_CATEGORY_FORMAT).concat(" ").concat(categoryNumeric);
-
-        Category category = null;
-        try {
-            category = guild.createCategory(categoryName).
-                    addRolePermissionOverride(
-                            guild.getPublicRole().getIdLong(),
-                            Collections.emptyList(),
-                            Collections.singletonList(Permission.VIEW_CHANNEL)).
-                    submit().get();
-
-            textCategories.add(category.getId());
-            settingsManager.set(DISCORDCHAT_TEXT_CATEGORY_IDS, textCategories);
-            settingsManager.save();
-        } catch (InterruptedException | ExecutionException ex) {
-            plugin.getLogger().log(Level.SEVERE, "Error while trying to create {0} category: " +
-                    ex.getMessage(), categoryName);
-        }
-
-        return category;
-    }
-    
     public void processChat(Source source, @NotNull Channel channel, @NotNull ClanPlayer clanPlayer, String message) {
         Clan clan = Objects.requireNonNull(clanPlayer.getClan(), "Clan cannot be null");
         List<ClanPlayer> receivers = new ArrayList<>();
@@ -193,16 +87,6 @@ public final class ChatManager {
         return parseChatFormat(format, message, new HashMap<>());
     }
 
-    public List<Long> getDiscordPlayersId(String clanTag) {
-        AccountLinkManager accountLinkManager = DiscordSRV.getPlugin().getAccountLinkManager();
-        return plugin.getClanManager().getClan(clanTag).getMembers().stream().
-                map(ClanPlayer::getUniqueId).
-                map(accountLinkManager::getDiscordId).
-                filter(Objects::nonNull).
-                map(Long::valueOf).
-                collect(Collectors.toList());
-    }
-
     public String parseChatFormat(String format, SCMessage message, Map<String, String> placeholders) {
         SettingsManager sm = plugin.getSettingsManager();
         ClanPlayer sender = message.getSender();
@@ -231,11 +115,6 @@ public final class ChatManager {
                 .replace("%message%", message.getContent());
     }
 
-    @NotNull
-    public Guild getGuild() {
-        return guild;
-    }
-
     private void registerHandlers() {
         Set<Class<? extends ChatHandler>> chatHandlers =
                 Helper.getSubTypesOf("net.sacredlabyrinth.phaed.simpleclans.chat.handlers", ChatHandler.class);
@@ -254,27 +133,11 @@ public final class ChatManager {
         }
     }
 
-    private List<ClanPlayer> getOnlineAllyMembers(Clan clan) {
-        return clan.getAllAllyMembers().stream().filter(allyPlayer -> allyPlayer.toPlayer() != null).collect(Collectors.toList());
+    public DiscordHook getDiscordHook() {
+        return discordHook;
     }
 
-    @NotNull
-    private List<String> getClanTags() {
-        List<String> whitelist = settingsManager.getStringList(DISCORDCHAT_TEXT_WHITELIST);
-
-        List<String> clanTags = plugin.getClanManager().getClans().stream().
-                filter(Clan::isVerified).
-                filter(clan -> !getDiscordPlayersId(clan.getTag()).isEmpty() || clan.isPermanent()).
-                map(Clan::getTag).
-                filter(clanTag -> whitelist.isEmpty() || whitelist.contains(clanTag)).
-                limit(settingsManager.getInt(DISCORDCHAT_TEXT_LIMIT)).
-                collect(Collectors.toList());
-
-        List<String> channels = guild.getTextChannels().stream().
-                map(GuildChannel::getName).
-                collect(Collectors.toList());
-
-        clanTags.removeAll(channels);
-        return clanTags;
+    private List<ClanPlayer> getOnlineAllyMembers(Clan clan) {
+        return clan.getAllAllyMembers().stream().filter(allyPlayer -> allyPlayer.toPlayer() != null).collect(Collectors.toList());
     }
 }
